requirements.txt
# DiffiT: Diffusion Vision Transformers for Image Generation
# Core dependencies

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
click>=8.0.0
psutil>=5.9.0
tqdm>=4.64.0

# Optional: faster PNG loading
# pyspng>=0.1.0

# Optional: TensorBoard logging
# tensorboard>=2.10.0

# Optional: Weights & Biases logging
# wandb>=0.13.0

# Development dependencies
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.10.0
# flake8>=5.0.0
